requirements.txt
# --- Interface et front-end ---
streamlit==1.39.0

# --- Manipulation de données ---
pandas==2.2.3
numpy==2.1.2
openpyxl

# --- Traitement et anonymisation de PDF ---
pdf2image==1.17.0
Pillow==10.4.0
PyMuPDF==1.24.10
opencv-python==4.10.0.84

# --- Traitement du texte ---
spacy==3.8.2
regex==2024.9.11

# --- Outils divers ---
python-docx==1.1.2
tqdm==4.66.5
